+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 -   Fixing unit type conversions when non-positional transforms are applied.
 -   Fixing variant propagation via `useAnimation()` when the parent component has no `variants` prop set.
 -   Fixing unsetting `whileHover` and `whileTap` if they contain `transitionEnd` values.
+-   Child components within variant trees now animate to `animate` as set by their parent.
+-   Checking animation props for array variants as well as strings.
 
 ## [1.2.4] 2019-07-15